finalizando validacoes, log

+++ backend/src/main/java/br/com/castgroup/cursos/controller/CursoController.java
@@ -4,6 +4,8 @@ import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 
+import javax.transaction.Transactional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.HttpStatus;
@@ -18,10 +20,13 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import br.com.castgroup.cursos.entities.Curso;
+import br.com.castgroup.cursos.entities.Log;
 import br.com.castgroup.cursos.form.FormCadastroCurso;
 import br.com.castgroup.cursos.form.FormUpdateCurso;
 import br.com.castgroup.cursos.repository.CategoriaRepository;
 import br.com.castgroup.cursos.repository.CursoRepository;
+import br.com.castgroup.cursos.repository.LogRepository;
+import br.com.castgroup.cursos.repository.UsuarioRepository;
 import br.com.castgroup.cursos.service.CursoService;
 
 @RestController
@@ -30,6 +35,12 @@ public class CursoController {
 
 	@Autowired
 	CursoService cursoService;
+	
+	@Autowired
+	UsuarioRepository usuarioRepository;
+	
+	@Autowired
+	LogRepository logRepository;
 
 	@Autowired
 	CursoRepository cursoRepository;
@@ -37,33 +48,35 @@ public class CursoController {
 	@Autowired
 	CategoriaRepository categoriaRepository;
 
+	@Transactional
 	@PostMapping
 	public ResponseEntity<String> cadastrar(@RequestBody FormCadastroCurso request) {
 		if (!cursoService.isValid(request, cursoRepository)) {
 			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(cursoService.mensagem());
 		}
 		Curso curso = request.converter(categoriaRepository);
+		Log log = new Log(null,LocalDate.now(),LocalDate.now(), curso, usuarioRepository.getById(1));
+		logRepository.save(log);
 		cursoRepository.save(curso);
 		return ResponseEntity.status(HttpStatus.CREATED).body(cursoService.mensagem());
 	}
 	
-	@SuppressWarnings("deprecation")
+	@Transactional
 	@PutMapping(value = "/{id_curso}")
 	public ResponseEntity<?> update(@PathVariable("id_curso") Integer id_curso, @RequestBody FormUpdateCurso request) {
 		try {
 			Optional<Curso> item = cursoRepository.findById(id_curso);
 			if (item.isEmpty()) {
 				return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Curso não encontrado");
-			} else {
-				
-				Curso curso = item.get();
-		
+			} else {				
+				Curso curso = item.get();		
 				if(!cursoService.isValid(request, cursoRepository, curso)) {
 					return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(cursoService.mensagem());
 				}
-				curso = request.converter(categoriaRepository, curso);
-			
+				curso = request.converter(categoriaRepository, curso);			
 				cursoRepository.save(curso);
+				Log log = new Log(null,curso.getInclusao(),LocalDate.now(), curso, usuarioRepository.getById(1));
+				logRepository.save(log);
 				return ResponseEntity.status(HttpStatus.OK).body("Curso atualizado");
 			}
 		} catch (Exception e) {
@@ -98,6 +111,7 @@ public class CursoController {
 		return cursoRepository.findByInicioBetween(inicio, termino);
 	}
 
+	@Transactional
 	@DeleteMapping(value = "/{id_curso}")
 	public ResponseEntity<String> deleteById(@PathVariable("id_curso") Integer id_curso) {
 		try {
@@ -106,6 +120,12 @@ public class CursoController {
 				return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Curso não encontrado");
 			} else {
 				Curso curso = item.get();
+				cursoService.finalizado(curso);
+				if (curso.getFinalizado()) {					
+					return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Curso finalizado não pode ser excluido");
+				}
+				Log log = new Log(null,curso.getInclusao(),LocalDate.now(), curso, usuarioRepository.getById(1));
+				logRepository.save(log);
 				cursoRepository.delete(curso);
 				return ResponseEntity.status(HttpStatus.OK).body("Curso deletado com sucesso");
 			}
